feat: add a Fetcher provider so it's always the same Fetcher Class instance
