+++ packages/react/src/hooks/useFetcher.ts
@@ -1,24 +1,6 @@
-import { useMemo } from "react";
+import { useContext } from "react";
 
 import { Fetcher } from "@carrot-kpi/sdk";
-import { useProvider } from "wagmi";
-import { usePreferDecentralization } from "./usePreferDecentralization";
-import { useIPFSGatewayURL } from "./useIPFSGatewayURL";
+import { FetcherContext } from "../providers";
 
-export function useFetcher() {
-    const preferDecentralization = usePreferDecentralization();
-    const ipfsGatewayURL = useIPFSGatewayURL();
-    const provider = useProvider();
-
-    const fetcher = useMemo(
-        () =>
-            new Fetcher({
-                provider,
-                ipfsGatewayURL,
-                preferDecentralization,
-            }),
-        [provider, ipfsGatewayURL, preferDecentralization]
-    );
-
-    return fetcher;
-}
+export const useFetcher = () => useContext<Fetcher>(FetcherContext);
