+++ packages/react/src/providers/index.ts
@@ -1,1 +1,2 @@
 export * from "./core-provider";
+export * from "./fetcher-provider";
